+++ client/components/views/auth.js
@@ -1,8 +1,8 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import { Switch, Route } from 'react-router-dom'
-import { auth } from '../store'
-import { Login, Signup } from '../components'
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { Switch, Route } from 'react-router-dom';
+import { auth } from '../../store';
+import { Login, Signup } from '../../components';
 
 class Auth extends Component {
   render() {
@@ -19,14 +19,14 @@ class Auth extends Component {
           </div>
         </div>
       </section>
-    )
+    );
   }
 }
 
 const mapState = state => ({
   isLoggedIn: !!state.me.isLoggedIn
-})
+});
 
-const mapDispatch = dispatch => ({})
+const mapDispatch = dispatch => ({});
 
-export default connect(mapState)(Auth)
+export default connect(mapState)(Auth);
