--- client/containers/index.js
@@ -1,1 +1,0 @@
-export { default as Auth } from './auth';
